LPS-189386 match sample
