+++ workspaces/liferay-ticket-workspace/client-extensions/liferay-ticket-etc-cron/Dockerfile
@@ -1,5 +1,5 @@
 FROM liferay/jar-runner:latest
 
-#ENV LIFERAY_JAR_RUNNER_JAVA_OPTS="-Xmx512m"
-
 COPY *.jar /opt/liferay/jar-runner.jar
+
+#ENV LIFERAY_JAR_RUNNER_JAVA_OPTS="-Xmx512m"
