--- workspaces/liferay-ticket-workspace/client-extensions/liferay-ticket-etc-cron/src/main/java/com/liferay/ticket/TicketCleanUpApplication.java
@@ -1,41 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.ticket;
-
-import com.liferay.client.extension.util.spring.boot.ClientExtensionUtilSpringBootComponentScan;
-
-import org.springframework.boot.WebApplicationType;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.context.annotation.Import;
-
-/**
- * @author Gregory Amerson
- */
-@Import(ClientExtensionUtilSpringBootComponentScan.class)
-@SpringBootApplication
-public class TicketCleanUpApplication {
-
-	public static void main(String[] args) {
-		new SpringApplicationBuilder(
-			TicketCleanUpApplication.class
-		).web(
-			WebApplicationType.NONE
-		).run(
-			args
-		);
-	}
-
-}
