+++ workspaces/liferay-ticket-workspace/client-extensions/liferay-ticket-etc-cron/src/main/java/com/liferay/ticket/TicketSpringBootApplication.java
@@ -1,0 +1,32 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.ticket;
+
+import com.liferay.client.extension.util.spring.boot.ClientExtensionUtilSpringBootComponentScan;
+
+import org.springframework.boot.WebApplicationType;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.context.annotation.Import;
+
+/**
+ * @author Gregory Amerson
+ */
+@Import(ClientExtensionUtilSpringBootComponentScan.class)
+@SpringBootApplication
+public class TicketSpringBootApplication {
+
+	public static void main(String[] args) {
+		new SpringApplicationBuilder(
+			TicketSpringBootApplication.class
+		).web(
+			WebApplicationType.NONE
+		).run(
+			args
+		);
+	}
+
+}
